requirements.txt
flask
flask-cors
flask-limiter
google-generativeai
python-dotenv
gunicorn